Enforce `int64` for `arange` in `_unique_internal`
Would be effectively enforced through copying the `arange` result to an
array with `int64` precision. Still it seems better to enforce this when
creating the array with `arange` to begin with.

+++ dask/array/routines.py
@@ -547,7 +547,7 @@ def _unique_internal(ar, indices, counts, return_inverse=False):
     r = np.empty(u.shape, dtype=dt)
     r["values"] = u
     if return_inverse:
-        r["inverse"] = np.arange(len(r))
+        r["inverse"] = np.arange(len(r), dtype=np.int64)
     for i, v in np.ndenumerate(u):
         if return_index:
             r["indices"][i] = indices[ar == v].min()
